Added the tabIndex call and a debug variable
Not sure if the tabIndex is really needed, as the lack of focus/receiving
keyboard events seems more to do with the Firefox dev console being open
and taking preference.

With the Firefox dev console closed, the canvas does seem to receive
keyboard events even without the tabIndex call.  Might as well leave it
in for now, just in case. :)