Use clip regions for what **might** be slightly better efficiency
Not sure if this will really be useful for this example, as the
current approach of having things drawn in a good order achieves
the same result.

With the clip region approach, there is a possibility of some
efficiency gain by the gpu not drawing unrendered areas.  I've not
even attempted measuring it though, so no idea if has any actual
effect or not.